[fixed] msvc compiling
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3193 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/faction/FactionManager.cpp
@@ -58,8 +58,8 @@ void FactionManager::awardFactionStanding(PlayerCreature* player, const String& 
 	if (!factionMap.contains(factionName))
 		return;
 
-	float lose = floor(130); //TODO: Find the equation for this.
-	float gain = floor(lose / 2); //you gain half of what you lose
+	float lose = floor((float)130); //TODO: Find the equation for this.
+	float gain = floor((float)lose / 2); //you gain half of what you lose
 
 	Faction faction = factionMap.get(factionName);
 	SortedVector<String>* enemies = faction.getEnemies();
