+++ MMOCoreORB/src/server/zone/managers/planet/MissionTargetMap.h
@@ -38,7 +38,7 @@ public:
 
 		float distance = 16000.f;
 		Coordinate coord(origin->getPositionX(), origin->getPositionZ(), origin->getPositionY());
-		coord.randomizePosition(1000 * pow(4, diff - 1));
+		coord.randomizePosition(1000 * pow((float)4, diff - 1));
 
 		if (coord.getPositionX() > zone->getMaxX())
 			coord.setPositionX(zone->getMaxX());
