+++ MMOCoreORB/src/server/zone/objects/creature/commands/EquilibriumCommand.h
@@ -75,7 +75,7 @@ public:
 			Time* cdTime = player->getCooldownTime("innate_equilibrium");
 
 			// Returns -time. Multiple by -1 to return positive.
-			int timeLeft = floor(cdTime->miliDifference() / 1000) *-1;
+			int timeLeft = floor((float)cdTime->miliDifference() / 1000) *-1;
 
 			stringId.setStringId("@innate:equil_wait"); // You are still recovering from your last equilization. Command available in %DI seconds.
 			stringId.setDI(timeLeft);
