+++ MMOCoreORB/src/server/zone/objects/structure/StructureObjectImplementation.cpp
@@ -218,7 +218,7 @@ String StructureObjectImplementation::getTimeString(uint32 timestamp) {
 	StringBuffer str;
 
 	for (int i = 3; i > -1; --i) {
-		values[i] = floor(timestamp / intervals[i]);
+		values[i] = floor((float)(timestamp / intervals[i]));
 		timestamp -= values[i] * intervals[i];
 
 		if (values[i] > 0) {
